feat: Split into files. Added detecting a disconnected monitor has returned
